requirements.txt
aiocsv==1.2.1
aiofiles==0.8.0
aiohttp==3.8.1
asyncio==3.4.3
autopep8==1.6.0
beautifulsoup4==4.10.0
black==22.1.0
flask==2.0.3
pip-chill==1.0.1
psycopg2-binary==2.9.3
pyopenssl==22.0.0
python-dotenv==0.19.2
pytz==2021.3
selenium==4.1.0
textblob==0.17.1
tweepy==4.5.0
